+++ index.js
@@ -1,1 +1,2 @@
-module.exports = require("./lib/iterate");
+require("@std/esm");
+module.exports = require("./lib/iterate.mjs").default;
