+++ lib/base-iterators/index.mjs
@@ -1,0 +1,9 @@
+/**
+ * @file Base utilities used to provide simple iteration of iterable objects
+ * @author Lachlan Hunt <lachlan.hunt@lachy.id.au>
+ */
+
+export const isIterable = require("./isiterable");
+export const iterate = require("./iterate");
+export const reverse = require("./reverse");
+export const safeIterate = require("./safeIterate");
