+++ lib/chainables/index.mjs
@@ -1,0 +1,30 @@
+/**
+ * @file Iterators that accept one or more iterators as input, and which may be chained.
+ * @author Lachlan Hunt <lachlan.hunt@lachy.id.au>
+ */
+
+export const chunk = require("./chunk");
+export const compact = require("./compact");
+export const concat = require("./concat");
+export const dedupe = require("./dedupe");
+export const difference = require("./difference");
+export const differenceBy = require("./differenceBy");
+export const differenceWith = require("./differenceWith");
+export const drop = require("./drop");
+export const dropUntil = require("./dropUntil");
+export const dropWhile = require("./dropWhile");
+export const everyNth = require("./everyNth");
+export const filter = require("./filter");
+export const flatten = require("./flatten");
+export const forEach = require("./forEach");
+export const fuzz = require("./fuzz");
+export const interlace = require("./interlace");
+export const link = require("./link");
+export const map = require("./map");
+export const modulo = require("./modulo");
+export const pluck = require("./pluck");
+export const take = require("./take");
+export const takeUntil = require("./takeUntil");
+export const takeWhile = require("./takeWhile");
+export const unzip = require("./unzip");
+export const zip = require("./zip");
