+++ lib/sequences/index.mjs
@@ -1,0 +1,15 @@
+export const { iterate, safeIterate, reverse } = require("../base-iterators");
+export const combination = require("./combination");
+export const counter = require("./counter");
+export const random = require("./random");
+export const range = require("./range");
+export const repeat = require("./repeat");
+export const permute = require("./permute");
+export const permutation = require("./permutation");
+export const series = require("./series");
+export const shuffle = require("./shuffle");
+export const symbols = require("./symbols");
+export const thueMorse = require("./thueMorse");
+export const collatz = require("./collatz");
+export const fibonacci = require("./fibonacci");
+export const fizzBuzz = require("./fizzBuzz");
